First version of the engine, with examples.
The first version of the engine has rudimentar scripting facilities,
support for circled shaped objects, one movement, movement limiter,
collision detection for circle objects, and the basics for developing
new objects and behaviors.

The scripting language will be kept simple, but a new grammar has
to be created, as the current one does not support property assign
or the execution of methods with parameters.