Add statement execution procedure to game interpreter.
Now the game interpreter can process expression, for `when` tests
in events), and statements for the `do` part in events.

To cope with the changes, all methods for behaviors should now
accept an arbitrary number of keyword arguments, and Game class
exposes a method to obtain a game object from its name.

The multi_ball example has changed to use the new implementation,
displaying better circle collision.